Tweets coming in are processed and saved to the DB
Right now this only works for single keyword searches, and doesn't check the database to see if there are word models that already exist. This will need to be worked on.

Also, this only logs the tweets and words in the database after the search is completed. Until then, they are stored in arrays in memory. This is due to the race conditions caused by the async operations needed to log them in the database. It is also significantly more effecient to save them all at once.

This can be mitigated by having a function on a setInterval that copies the words array, empties it, and then saves the copied array to the database. This would allow for progressive updating of the database but possibly prevent the race conditions in the original solution.