Tweets are no longer saved.
Saving Tweets proved to be fairly time intensive, and when retrieving even a small dataset, caused the retrieval to take a long time. Removing these led to a large decrease in the time for retrieval using the API.

I still need to remove the processing of the full tweets for extra performance savings.